fix(drive): keep drive-proof-verifier building drive with verify only
The tenderdash trust-boundary tests dev-depended on drive's full
feature from inside the client proof-verifier crate; feature
unification then compiled node-side storage/proving into the client
crate's test build, violating the client/server dependency boundary.
Move the suite to rs-drive-abci (which already builds drive server and
already dev-depends on drive-proof-verifier) as having_trust_boundary:
same four tests, now generating proofs on the server side and running
the client wrapper against them. drive-proof-verifier's drive
dependency is back to verify-only in both sections.
